patch: moves `base58` and `base64` into `encoding`
